change to long

+++ features/transaction/cart/src/main/java/com/tokopedia/cart/bundle/view/CartListPresenter.kt
@@ -466,7 +466,7 @@ class CartListPresenter @Inject constructor(private val getCartRevampV3UseCase: 
                     )
                 }.toList()
         )
-        summaryTransactionUiModel?.promoValue = lastApplyUiModel.benefitSummaryInfo.finalBenefitAmount
+        summaryTransactionUiModel?.promoValue = lastApplyUiModel.benefitSummaryInfo.finalBenefitAmount.toLong()
     }
 
     override fun reCalculateSubTotal(dataList: List<CartShopHolderData>) {
@@ -493,13 +493,13 @@ class CartListPresenter @Inject constructor(private val getCartRevampV3UseCase: 
     private fun updateSummaryTransactionUiModel(subtotalBeforeSlashedPrice: Double, subtotalPrice: Double, totalItemQty: Int, subtotalCashback: Double) {
         summaryTransactionUiModel?.qty = totalItemQty.toString()
         if (subtotalBeforeSlashedPrice == 0.0) {
-            summaryTransactionUiModel?.totalValue = subtotalPrice.toInt()
+            summaryTransactionUiModel?.totalValue = subtotalPrice.toLong()
         } else {
-            summaryTransactionUiModel?.totalValue = subtotalBeforeSlashedPrice.toInt()
+            summaryTransactionUiModel?.totalValue = subtotalBeforeSlashedPrice.toLong()
         }
-        summaryTransactionUiModel?.discountValue = (subtotalBeforeSlashedPrice - subtotalPrice).toInt()
-        summaryTransactionUiModel?.paymentTotal = subtotalPrice.toInt()
-        summaryTransactionUiModel?.sellerCashbackValue = subtotalCashback.toInt()
+        summaryTransactionUiModel?.discountValue = (subtotalBeforeSlashedPrice - subtotalPrice).toLong()
+        summaryTransactionUiModel?.paymentTotal = subtotalPrice.toLong()
+        summaryTransactionUiModel?.sellerCashbackValue = subtotalCashback.toLong()
     }
 
     private fun getAvailableCartItemDataList(dataList: List<CartShopHolderData>): ArrayList<CartItemHolderData> {
